fix(llma): apply caching costs to all providers
